--- src/Sylius/Bundle/CoreBundle/Resources/config/services/payment_request/payum/command_handlers.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Sylius Sp. z o.o.
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
->
-    <services>
-        <defaults public="true" />
-
-        <service id="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\TokenPaymentRequestHandler" abstract="true">
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
-            <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
-            <argument type="service" id="sylius.payum.payment_request.factory.payum_token" />
-            <argument type="service" id="sylius.payum.payment_request.processor.request" />
-            <argument type="service" id="sylius.payum.payment_request.processor.after_token_request" />
-        </service>
-
-        <service id="sylius.payment_request.payum.capture" parent="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\TokenPaymentRequestHandler">
-            <argument type="service" id="sylius.factory.payum_capture" />
-            <tag name="messenger.message_handler" bus="sylius.payment_request.command_bus" handles="Sylius\Bundle\CoreBundle\PaymentRequest\Command\Payum\CapturePaymentRequest" />
-        </service>
-
-        <service id="sylius.payment_request.payum.authorize" parent="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\TokenPaymentRequestHandler">
-            <argument type="service" id="sylius.factory.payum_authorize" />
-            <tag name="messenger.message_handler" bus="sylius.payment_request.command_bus" handles="Sylius\Bundle\CoreBundle\PaymentRequest\Command\Payum\AuthorizePaymentRequest" />
-        </service>
-
-        <service id="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\ModelPaymentRequestHandler" abstract="true">
-            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
-            <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
-            <argument type="service" id="sylius.payum.payment_request.processor.request" />
-        </service>
-
-        <service id="sylius.payment_request.payum.status" parent="Sylius\Bundle\CoreBundle\PaymentRequest\CommandHandler\Payum\ModelPaymentRequestHandler">
-            <argument type="service" id="sylius.factory.payum_get_status_action" />
-            <tag name="messenger.message_handler" bus="sylius.payment_request.command_bus" handles="Sylius\Bundle\CoreBundle\PaymentRequest\Command\Payum\StatusPaymentRequest" />
-        </service>
-
-    </services>
-</container>
