+++ src/Sylius/Bundle/PayumBundle/PaymentRequest/Command/AuthorizePaymentRequest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\Command\Payum;
+namespace Sylius\Bundle\PayumBundle\PaymentRequest\Command;
 
 use Sylius\Bundle\PaymentBundle\Command\PaymentRequestHashAwareInterface;
 use Sylius\Bundle\PaymentBundle\Command\PaymentRequestHashAwareTrait;
@@ -20,8 +20,7 @@ class AuthorizePaymentRequest implements PaymentRequestHashAwareInterface
 {
     use PaymentRequestHashAwareTrait;
 
-    public function __construct(
-        protected ?string $hash,
-    ) {
+    public function __construct(protected ?string $hash)
+    {
     }
 }
