+++ src/Sylius/Bundle/PayumBundle/Resources/config/services/payment_request/command_handler.xml
@@ -1,0 +1,51 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container
+    xmlns="http://symfony.com/schema/dic/services"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
+    <services>
+        <defaults public="true" />
+
+        <service id="sylius.payum.payment_request.command_handler.token_payment_request" class="Sylius\Bundle\PayumBundle\PaymentRequest\CommandHandler\TokenPaymentRequestHandler" abstract="true">
+            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
+            <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
+            <argument type="service" id="sylius.payum.payment_request.factory.payum_token" />
+            <argument type="service" id="sylius.payum.payment_request.processor.request" />
+            <argument type="service" id="sylius.payum.payment_request.processor.after_token_request" />
+        </service>
+
+        <service id="sylius.payum.payment_request.command_handler.token_payment_request.capture" parent="sylius.payum.payment_request.command_handler.token_payment_request">
+            <argument type="service" id="sylius.factory.payum_capture" />
+            <tag name="messenger.message_handler" bus="sylius.payment_request.command_bus" handles="Sylius\Bundle\PayumBundle\PaymentRequest\Command\CapturePaymentRequest" />
+        </service>
+
+        <service id="sylius.payum.payment_request.command_handler.token_payment_request.authorize" parent="sylius.payum.payment_request.command_handler.token_payment_request">
+            <argument type="service" id="sylius.factory.payum_authorize" />
+            <tag name="messenger.message_handler" bus="sylius.payment_request.command_bus" handles="Sylius\Bundle\PayumBundle\PaymentRequest\Command\AuthorizePaymentRequest" />
+        </service>
+
+        <service id="sylius.payum.payment_request.command_handler.model_payment_request" class="Sylius\Bundle\PayumBundle\PaymentRequest\CommandHandler\ModelPaymentRequestHandler" abstract="true">
+            <argument type="service" id="Sylius\Bundle\CoreBundle\PaymentRequest\Provider\PaymentRequestProviderInterface" />
+            <argument type="service" id="sylius.payum.payment_request.resolver.doctrine_proxy_object_resolver" />
+            <argument type="service" id="sylius.payum.payment_request.processor.request" />
+        </service>
+
+        <service id="sylius.payum.payment_request.command_handler.model_payment_request.status" parent="sylius.payum.payment_request.command_handler.model_payment_request">
+            <argument type="service" id="sylius.factory.payum_get_status_action" />
+            <tag name="messenger.message_handler" bus="sylius.payment_request.command_bus" handles="Sylius\Bundle\PayumBundle\PaymentRequest\Command\StatusPaymentRequest" />
+        </service>
+    </services>
+</container>
